Move map screen into a specific component
